+++ src/main/scala/dev/krysztal/are/common/Registries.scala
@@ -22,39 +22,19 @@
  * SOFTWARE.
  */
 
-package dev.krysztal.are
+package dev.krysztal.are.common
 
-import dev.krysztal.are.common.Blocks
-import dev.krysztal.are.common.ChunkComponents
-import dev.krysztal.are.common.DataComponentTypes
-import dev.krysztal.are.common.ItemGroups
-import dev.krysztal.are.common.Items
-import dev.krysztal.are.common.Registries
-import dev.krysztal.are.common.WorldComponents
-import net.fabricmc.api.ModInitializer
-import net.minecraft.util.Identifier
-import org.slf4j.LoggerFactory
+import dev.krysztal.are.common.gene.GeneticSequence
+import net.fabricmc.fabric.api.event.registry.FabricRegistryBuilder
+import net.fabricmc.fabric.api.event.registry.RegistryAttribute
+import net.minecraft.registry.SimpleRegistry
 
-class AgrarianRevolution extends ModInitializer {
+object Registries {
+    val GENETIC_SEQUENCE: SimpleRegistry[GeneticSequence] =
+        FabricRegistryBuilder
+            .createSimple(RegistryKeys.GENETIC_SEQUENCE)
+            .attribute(RegistryAttribute.SYNCED)
+            .buildAndRegister()
 
-    def onInitialize() = {
-        Registries.initialize()
-
-        DataComponentTypes.initialize()
-
-        WorldComponents.initialize()
-        ChunkComponents.initialize()
-
-        ItemGroups.initialize()
-        Items.initialize()
-        Blocks.initialize()
-    }
-}
-
-object AgrarianRevolution {
-    private lazy val log = LoggerFactory.getLogger("AgrarianRevolution")
-
-    val modID = "are"
-
-    def identifier(id: String) = Identifier.of(modID, id)
+    def initialize() = {}
 }